6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,8 +10,8 @@ int main(void)
 {
 	printf("size of a char: %lu bytes\n", sizeof(char));
 	printf("size of an int: %lu bytes\n", sizeof(int));
-	printf("size of a long int: %lu bytes\n", sizeof(long));
-	printf("size of a long long int: %lu bytes\n", sizeof(double));
+	printf("size of a long int: %lu bytes\n", sizeof(long int));
+	printf("size of a long long int: %lu bytes\n", sizeof(long long int));
 	printf("size of a float: %lu bytes\n", sizeof(float));
